requirements.txt
streamlit>=1.32.0
langchain>=0.2.14
langchain-core>=0.2.0
langchain-google-genai>=1.0.7
langgraph>=0.2.0
google-generativeai>=0.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0
